Rename `core` to `uri`
