+++ src/mod.ts
@@ -2,8 +2,8 @@
  * @module @hatsy/churi
  */
 export * from './charge/mod.js';
-export * from './core/mod.js';
 export * from './rx/mod.js';
 export * from './schema/mod.js';
 export * from './syntax/mod.js';
 export * from './uncharge/mod.js';
+export * from './uri/mod.js';
